deniz

+++ agro_alert_web/app/GraphQl/Tespit.graphql.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const GET_TESPITLER_BY_GOZLEM_ID = gql`
-query getTespitlerByGozlemId($gozlemId: Float!) {
+query getTespitlerByGozlemId($gozlemId: Int!) {
   tespitlerByGozlemId(gozlemId: $gozlemId) {
     id
     tespit_tipi
@@ -14,7 +14,7 @@ query getTespitlerByGozlemId($gozlemId: Float!) {
 `;
 
 export const GET_TESPIT_BY_ID = gql`
-query getTespitById($id: Float!) {
+query getTespitById($id: Int!) {
   tespitById(id: $id) {
     id
     tespit_tipi
